chore: move txtype to dedicated mod
